+++ src/query/expression/src/kernels/mod.rs
@@ -21,7 +21,6 @@ mod sort;
 mod take;
 mod take_chunks;
 mod take_compact;
-mod tests;
 mod topk;
 mod utils;
